requirements.txt
flask==2.3.2
flask-cors==4.0.0
transformers==4.28.0
torch==2.0.0
google-generativeai==0.3.2
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.31.0